+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Removed fuel & payload selector (cannot be set in Aerofly FS)
 - Added correct UID generator via `@fboes/aerofly-custom-missions`
 - Disable METAR button in GUI if date is out of range
+- Changed MCP capability for setting intial aircraft state / speed
 
 ## [2.0.2] - 2026-07-19